pfkhra si/backend/authentication/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import (
    login_view, logout_view, current_user_view, get_csrf_token,
    UserViewSet, AuditLogViewSet
)

router = DefaultRouter()
router.register(r'users', UserViewSet, basename='user')
router.register(r'audit-logs', AuditLogViewSet, basename='auditlog')

urlpatterns = [
    # Authentication endpoints
    path('auth/login/', login_view, name='login'),
    path('auth/logout/', logout_view, name='logout'),
    path('auth/me/', current_user_view, name='current-user'),
    path('auth/csrf/', get_csrf_token, name='csrf-token'),
    
    # Router endpoints (users, audit-logs)
    path('', include(router.urls)),
]
